+++ src/app/shared/PersonLikes.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+  import {seconds} from "hurdak"
   import Feed from "src/app/shared/Feed.svelte"
 
   export let pubkey
@@ -7,4 +8,4 @@
   const filter = {kinds: [7], authors: [pubkey]}
 </script>
 
-<Feed hideControls {relays} {filter} />
+<Feed hideControls delta={seconds(12, "hour")} {relays} {filter} />