+++ src/app/shared/PersonNotes.svelte
@@ -1,6 +1,8 @@
 <script lang="ts">
+  import {seconds} from "hurdak"
   import {noteKinds} from "src/util/nostr"
   import Feed from "src/app/shared/Feed.svelte"
+  import {sessions} from "src/engine2/state"
 
   export let pubkey
   export let relays
@@ -8,4 +10,9 @@
   const filter = {kinds: noteKinds, authors: [pubkey]}
 </script>
 
-<Feed {relays} {filter} invertColors />
+<Feed
+  invertColors
+  delta={seconds(12, "hour")}
+  queryCache={Boolean($sessions[pubkey])}
+  {relays}
+  {filter} />