+++ src/util/nostr.ts
@@ -147,7 +147,7 @@ export const isShareableRelay = (url: string) =>
 
 export const normalizeRelayUrl = (url: string) => {
   // If it doesn't start with a compatible protocol, strip the proto and add wss
-  if (!url.match(/^wss:\/\/.+/)) {
+  if (!url.match(/^(wss|local):\/\/.+/)) {
     url = "wss://" + url.replace(/.*:\/\//, "")
   }