+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - [x] Add ToS and Privacy Policy
 - [x] Use an ephemeral key when reporting content and expire reports
 - [x] Fix message ordering
+- [x] Add local relay
+- [x] Reduce wait time for person feeds
 
 # 0.3.6